removed some superfluous function args, code does not compile
